Use real device info.
kernel.diff is the required kernel changes (only uaudio and azalia
implement the required logic for now).

Currently, only a small display name is possible (i.e. they get
truncated), due to ABI constraints in both the kernel and in userspace.
When we are further down the line, we can break ABI compat.

Example:
```
$ sndioctl -d
...
015:server.device=0(Realtek ALC)
020:server.device=1(Logitech HD)
021:server.device=2(SteelSeries)
```